application/requirements.txt
fdb
python-telegram-bot
python-telegram-bot[job-queue]
python-dotenv
psycopg2

#anyio==4.0.0
#APScheduler==3.10.4
#certifi==2023.7.22
#exceptiongroup==1.1.3
#fdb==2.0.2
#future==0.18.3
#h11==0.14.0
#httpcore==0.17.3
#httpx==0.24.1
#idna==3.4
#psycopg2==2.9.8
#python-dotenv==1.0.0
#python-telegram-bot==20.5
#pytz==2023.3.post1
#six==1.16.0
#sniffio==1.3.0
#tzlocal==5.0.1
